INVESTIGATE: MSVC ffi
Unit tests use the lua ffi to call Neovim C code, in MSVC this currently
broken:

1. The preprocess.lua module we use to do C preprocessing does not handle
   Windows/MSVC
2. The luajit ffi does not handle the following
   - MSVC linemarkers e.g #line 1
   - __pragma extension
3. For some reason some C header have formfeed characters in there, as in
   \f\n
4. preprocess.lua has some subtle bugs, e.g. the build path is hardcoded, and
   its basically repeating the same detection work done in CMake

This commit handles 1, 3 and 4 - my goal was to completely move the preprocessor
calls out of the lua module and have CMake call it. But then I got stuck in the
__pragma handle since there are non trivial code blocks in the headers.

I'm keeping this out of the tree, since it like breaks UNIX tests as is.