Link against winpty.dll.
WINPTY_CHECKOUT_PATH needs to defined on the cmake command-line.  e.g.
Using -DWINPTY_CHECKOUT_PATH=<path>.

The winpty checkout needs to be compiled (using any Cygwin or MSYS
environment), using the same 32-bit or 64-bit architecture as neovim.

This integration is intended to work with the neovim-demo branch of
https://github.com/rprichard/winpty.git.